updated start and stop scripts

+++ stop.sh
@@ -1,2 +1,9 @@
-kubectl delete pod spring-boot-kubernates
-kill -9 $(lsof -t -i:${PORT})
+PID_FILE=spring-boot-portforward.pid
+if [[ -f "$PID_FILE" ]]; then
+    echo "Stopping"
+
+	kill -9 $(cat ${PID_FILE})
+	rm ${PID_FILE}
+
+	kubectl delete pod spring-boot-kubernates
+fi
